Crate vswhom

FFI to Jon Blow's VS discovery script.

See the associated funxions on VsFindResult for more specific info, but I'll copy over some usage information from the original file:

The purpose of this file is to find the folders that contain libraries you may need to link against, on Windows, if you are linking with any compiled C or C++ code. This will be necessary for many non-C++ programming language environments that want to provide compatibility.

We find the place where the Visual Studio libraries live (for example, libvcruntime.lib), where the linker and compiler executables live (for example, link.exe), and where the Windows SDK libraries reside (kernel32.lib, libucrt.lib).

One other shortcut I took is that this is hardcoded to return the folders for x64 libraries. If you want x86 or arm, you can make slight edits to the code below, or, if enough people want this, I can work it in here.

Structs

VsFindResult
The result of looking for Visual Studio and Windows SDK.